refactor!: Move bundled extensions into mopidy._exts
