+++ src/mopidy/_exts/http/data/clients.html
