+++ src/mopidy/_exts/http/data/favicon.ico
