+++ src/mopidy/_exts/http/data/mopidy.css
