+++ src/mopidy/_exts/m3u/ext.conf
