+++ tests/_exts/file/conftest.py
@@ -2,7 +2,7 @@ from unittest import mock
 
 import pytest
 
-from mopidy.file import backend
+from mopidy._exts.file import backend
 from tests import path_to_data_dir
