+++ tests/_exts/http/test_events.py
@@ -2,11 +2,11 @@ import json
 
 from pytest_mock import MockerFixture
 
-from mopidy.http import actor
+from mopidy._exts.http import actor
 
 
 def test_track_playback_paused_is_broadcasted(mocker: MockerFixture) -> None:
-    broadcast = mocker.patch("mopidy.http.handlers.WebSocketHandler.broadcast")
+    broadcast = mocker.patch("mopidy._exts.http.handlers.WebSocketHandler.broadcast")
     io_loop = mocker.Mock()
 
     actor.on_event("track_playback_paused", io_loop, foo="bar")
@@ -18,7 +18,7 @@ def test_track_playback_paused_is_broadcasted(mocker: MockerFixture) -> None:
 
 
 def test_track_playback_resumed_is_broadcasted(mocker: MockerFixture) -> None:
-    broadcast = mocker.patch("mopidy.http.handlers.WebSocketHandler.broadcast")
+    broadcast = mocker.patch("mopidy._exts.http.handlers.WebSocketHandler.broadcast")
     io_loop = mocker.Mock()
 
     actor.on_event("track_playback_resumed", io_loop, foo="bar")
